- Tests for Movie API finished

+++ api/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 class Movie(models.Model):
     name = models.CharField(max_length=150)
-    year_of_release = models.IntegerField()
+    year_of_release = models.PositiveSmallIntegerField()
 
     def __str__(self):
         return self.name
